refactor delete rom dialog as component
